Consolidate admin components into components/admin directory
