+++ src/components/admin/AccountFormHeader.js
@@ -1,3 +1,5 @@
+import React from "react"
+
 const AccountFormHeader = ({ title, subtitle }) => (
 	<span className="text-center dark-green-text">
 		<h2 className="text-uppercase page-title">{title}</h2>
@@ -5,7 +7,7 @@ const AccountFormHeader = ({ title, subtitle }) => (
 		<style jsx>{`
 			.page-title {
 				font-family: 'Open Sans Condensed', sans-serif;
-				font-size: 60px;
+				font-size: 45px;
 				text-transform: uppercase;
 			}
