+++ src/pages/admin/login.js
@@ -1,8 +1,8 @@
 import React from "react"
 import Link from "next/link"
 import Layout from "../../components/Layout"
-import AccountFormHeader from "../../components/account/AccountFormHeader"
-import AccountForm from "../../components/account/AccountForm"
+import AccountFormHeader from "../../components/admin/AccountFormHeader"
+import AccountForm from "../../components/admin/AccountForm"
 
 const Login = () => (
 	<Layout>
@@ -29,7 +29,7 @@ const Login = () => (
 		</span>
 		<style jsx>{`
 			.responsive-width {
-				width: 40%;
+				width: 33%;
 			}
 			
 			.custom-border {
@@ -59,7 +59,7 @@ const Login = () => (
 
 			.page-title {
 				font-family: 'Oswald', sans-serif;
-				font-size: 60px;
+				font-size: 45px;
 			}
 
 			.dark-green-text, .dark-green-text a {
